Move plotting function from make-plots.py to dctools/plot.py to permit importing and using in notebooks
